requirements.txt
pymongo==4.6.1
colorama==0.4.6
mysql-connector-python==8.0.33
